use 2-timers approach and adjust tests
