+++ packages/utils/src/index.ts
@@ -27,4 +27,3 @@ export * from './clientreport';
 export * from './ratelimit';
 export * from './baggage';
 export * from './url';
-export * from './debounce';
